+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "vsc-webshark" extension will be documented in this file.
 
+## [0.1.0]
+- time sync part 1: double click broadcasts the (utc) time from the selected frame.
+- filter are persisted on vscode restart
+
 ## [0.0.2]
 - Close sharkd child process on closing the document
 - document retains content when put in the background